fix: Warn if `CAP_IPC_LOCK` capability is missing
